[#11] refactor: primitive type -> reference type 및 dto 디렉토리 경로 수정
