+++ src/main/java/com/dnd5th3/dnd5th3backend/controller/MemberController.java
@@ -1,7 +1,6 @@
 package com.dnd5th3.dnd5th3backend.controller;
 
-import com.dnd5th3.dnd5th3backend.controller.dto.MemberRequestDto;
-import com.dnd5th3.dnd5th3backend.repository.MemberRepository;
+import com.dnd5th3.dnd5th3backend.controller.dto.member.MemberRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
