+++ src/main/java/com/dnd5th3/dnd5th3backend/controller/dto/member/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package com.dnd5th3.dnd5th3backend.controller.dto;
+package com.dnd5th3.dnd5th3backend.controller.dto.member;
 
 import com.dnd5th3.dnd5th3backend.domain.member.Member;
 import com.dnd5th3.dnd5th3backend.domain.member.Role;
